requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-Mail==0.9.1
gunicorn==22.0.0
requests==2.32.3
openai==1.42.0
python-dotenv==1.0.1
